its a mess
currently in process of copying everything in the class from the _app.py to make seperate State 1 and State 2 calculations. im going to have to go through and change all self.variables and functionnames into self.S1variables and S1functionname and self.S2variables and S2functionname. i think i have everything needed for the seperate state calculations seperated out, there is a line, "# state 2 starts here" line 830, i believe everything needed to calculate the state 2 is under that line.  

currently state 2 is the only one that is calculated as i left state 2 as the original, state 1 is in the process of everything being labeled as S1blahblah, so it is not currently being calculated.