Halfway there
It now calculates both states and displays them, it needs to be made into a horizontal display per the HW7doc but that should be easy enough to figure out. So far everything labeled S1xxxxxx is for state 1, i didnt label the state 2 variables and functions and i dont think its worth doing because it was tedious to trace down all variables for State 1 that needed to be labeled and they should all be seperated(S1 for state 1, no label for state 2) now anyways.